tournament/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filetype==1.2.0
idna==3.10
Pillow==10.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
